Add the Agify API
This commit adds the Agify API to get the median age of
a certain name in a certain country. It adds a client used
in conjuction with Retrofit to get the age of a certain name.
It uses AgifyUser to model that data as a Java object.
This commit also adds the IT for the client and the test for the
data model.