Add the Agify service and the cache
This commit adds the Agify service to our web application.
It also adds the unit test for this service. It also adds the
IT for the application (MeetifyControllerIT).
This commit adds the test for UserData. It fixes an issue
with the output from Agify by invoking Jackson to convert
said output. It also adds random improvements like
a user ID for the standard user and the Agify user.
This commit also adds the cache feature in our program.